Make ObjectType a base class
ObjectType is now a base class that can be subclassed for specialized purposes
so that, for example, manager type objects need not implement the full
capabilities needed by in-game objects (now implemented with the
CollideableObjectType subclass), such as collision detection.
Object YAML files now require higher-level hashes, keyed by the name of the
ObjectType subclass that the list(s) of defined objects will be instances of.
The overall goal is to make it possible for game developers to subclass their
own object types, and to enable the creation of GUI widgets and containers.
Updated test_object_type.py.
Updated demo and test YAML object lists.
Added a SimpleObjectInstance class with a subset of ObjectInstance capabilities
that ObjectTypes can manage, that lacks complicated sprite handling.
Looking into making SimpleObjectInstance a base class for ObjectInstance.