Color class now wraps pygame.Color
pygame.Color is almost good enough, but is missing a number of CSS color names,
as well as the ability to specify lists as a parameter.  Spell out individual
color components as color.red, color.green, color.blue and color.alpha, and
retrieve a color component list using color.rgb or color.rgba.
An attempt to subclass pygame.Color failed due to it bombing out of the script
with a ValueError when specifying an unknown color, even though __init__()
was overridden and was calling super() inside a try / except block.  Debug
output from the subclass __init__() wouldn't even print before the super()
call.