Fix object type scheduled deletion bug
Scheduled instance deletion in the base class was calling the remove() list
method with a list as the argument.  This is ok for the collideable object
subclass, since the group attribute remove() method is capable of acting on an
iterable, but is an error for the base class.  Delete the items in a loop
instead.