ObjectInstance: Saner child instance bounds detection
The root instance handles updates for all its descendants.  This makes it
possible to queue all bounds events for a tree of instances and only transmit
each event once.
Also, fixed a boundary detection bug in which child instances in the
test_object_type.py unit test would detect outside_room events, but the event
never got queued so the event engine would raise an exception when the event
was transmitted.