ObjectSprite: Initial support for image strips
An ObjectSprite given a filename ending in _strip#.EXT, where '#' indicates
how many sub-images are inside the image, and EXT is the image file type,
will generate that many subimages when load_graphic() is called.

Custom left columns may be specified in the "custom_subimage_columns" keyword
argument in the constructor (or hash key in a YAML description), which should
be passed as a list with its length matching the subimage count.  Any extra list
elements will be dropped, and an attempt will be made to fill in missing list
elements evenly divided between any remaining space in the image.
By default, left columns will be calculated so that each subimage has the same
width, and all subimages fit within the full image width.

The basic unit test passes. The custom_subimage_columns are not yet tested.