Add font resource and unit test
In preparation for drawing text, the new Font class wraps font resources.
Fonts can be different sizes, and support bold, italic, underline and anti-
aliasing effects.  A line spacing can also be specified, which is used when
rendering multi-line text.
The new FontRenderer class is returned from the Font.get_font_renderer() method,
and is the object used to draw multi-line text on a surface.  Call the
FontRenderer.calc_render_size() method with the text to be drawn to get the
full width and height for the block of text.  Call the
FontRenderer.render_text() method with the surface to draw to, the
newline-delimited text string, a coordinate, and foreground color (and
optionally background color) to draw the text block.  If no background color is
set, the text will be drawn to a transparent background.