Font unit test: Add font selection table
Fonts can now be navigated with the arrow keys.  Moving up from the top row
will go to the previous page of fonts (if not at page 0), and moving down from
the bottom row will go to the next page of fonts (if not at the last page).
Moving left from the left column will move to the rightmost column, and moving
right from the right column will move to the leftmost column.
The text drawing is now in a separate method called from draw_objects().