Font: Support load from file. Add caching.
The Font class now looks for a 'file://' prefix on fontname, and if found,
uses the remainder of the string as the path to a .ttf file containing a
desired font.  Another '/' immediately after file:// indicates a full path.
If bold or italic flags are set, the pygame.font methods set_bold() and
set_italic() will be applied, though it is better to load multiple font files
each with the desired effects.
Font renderers are now cached for re-use.  Rendered text is now cached as
well (each line is cached separately).  This can increase the memory usage
without any bound at this time.  May be necessary to create a FontCache class
that drops rarely used and/or old entries to mitigate excessive memory usage.
The unit test now draws a rectangle around the source code display, to show
how accurate the size calculations are for each font.
Added the DroidSansFallbackFull.ttf file used in the unit test for font file
loading.