Fixed widget style bug
Widgets were overriding their symbol table entries for CSS properties,
instead of the other way around.
Since the CSS properties now contain the complete list, there's no need
for default symbol table entries. This change also removed a bug in which
a color was stored in the symbol table as a Color object instead of a
string.