Adjust resource load paths
Backgrounds weren't loading from the backgrounds/ path.  The game engine
had a bug in the directory name ('background' instead of 'backgrounds'),
and also backgrounds were loading graphics whenever commanded instead of
caching the image loading (so setup() wasn't useful), and the current
directory was not backgrounds/ at the time the image was loaded.

Also changed the Sound class attribute 'sound_file' to 'filename' to
match other resources (and so the path adjustment method can prepend
'sounds/' to the file name).