python3 support: better fix for Coordinate object comparison
The problem encountered in test code was that position attributes
were initially coming in as (x, y) typle but got stored as a
Coordinate object, and python 3 started using Coordinate.__eq__()
to do the comparisons.

My original approach was to try making the new position also a
Coordinate before plugging into the compare, so __eq__() would
know what to do with it.

This had an invalid assumption though, that new_val / old_val
would always be coordinates... actually they are not. They are
general attributes, which are sometimes of non-coordinate types.

So instead of forcing new_val into a Coordinate before doing the
comparison, do any tampering with the type inside Coordinate.__eq__()
method -- where a second Coordinate is always wanted. Other types of
attributes will use their own respective __eq__() methods and won't
end up in the Coordinate one.