language_engine.py: Reduce redundant debug; fix exception syntax errors
Printing the parsestr parameter in self.debug() results in the same
string printed multiple times, once for each token in a string.  Remove
it to make the debug output easier to manage.

Fixed a few cases where parameters to exceptions ended up in format
strings instead (and cause "too many arguments for format string" errors).

"import operator" is needed for eval(), so tell pylint to ignore it.