New automated_tests.sh script
This script will run all non-interactive tests in pygame/tests (assuming
PYTHONPATH is set ahead of time, or the engine has been installed on the
system).
Any test scripts that failed will be summarized at the end.
Also, fixed mouse and keyboard issues in test_room.py caused by changes
to the keyboard event naming and game engine execute_action parameters
(these events were not needed for the test, but would stop the test code
with an exception when generated).